New Area Text
Added NewAreaTextTrigger prefab to prefabs/UI. To use it, drag the prefab into the scene and change its color and text fields as desired.

Displays animated text on screen when entering a new area

Can be set to be repeatable. If repeatable, cannot be triggered again until the animated text prefab sends a signal to the trigger that it's done animating.